Make maintenance signoff evidence reproducible
Split the BV validator monolith into focused modules so the critical file-bloat finding becomes an accounted follow-up set instead of hiding inside one oversized script. Also update QC selector references to the active-review lane that actually drives current maintenance signoff.

Constraint: maintenance-code-health remains the only release signoff gap until repo-wide follow-up debt is either burned down or explicitly accepted.

Rejected: leaving the split modules ignored and unstaged | local validation would not reproduce on CI or another checkout.

Confidence: high

Scope-risk: moderate

Directive: keep validate:bv parity as the guardrail for future BV validator extractions; do not flip maintenance-code-health to ready-with-scope without an explicit follow-up debt decision.

Tested: npm.cmd run verify:qc; npm.cmd run verify:rules; npm.cmd run electron:test:build; npm.cmd run validate:bv; npm.cmd run typecheck -- --pretty false; npm.cmd run qc:select -- --status=active-review

Not-tested: full verify:app-completion:release still expected to fail while maintenance-code-health is active-review.
