+++ scripts/validate-bv-known-units.ts
@@ -1,0 +1,104 @@
+export const CLAN_CHASSIS_MIXED_UNITS = new Set([
+  'alpha-wolf-a',
+  'alpha-wolf-c',
+  'alpha-wolf-prime',
+  'archer-arc-7c',
+  'atlas-c',
+  'atlas-ii-as7-dk-h',
+  'atlas-iii-as7-d2',
+  'atlas-iii-as7-d3',
+  'avalanche-avl-1or',
+  'avatar-av1-or',
+  'bakeneko-bkn-1k',
+  'black-hawk-ku-bhku-or',
+  'black-hawk-r',
+  'black-hawk-u',
+  'black-hawk-x',
+  'black-lanner-i',
+  'blackjack-bj2-or',
+  'bull-shark-bsk-maz',
+  'charger-c',
+  'cougar-xr',
+  'deimos-e',
+  'dervish-dv-11dk',
+  'devastator-dvs-2-ec',
+  'dragonfly-r',
+  'dragonfly-u',
+  'enfield-end-6j-ec',
+  'epimetheus-prime',
+  'fenris-i',
+  'fire-falcon-r',
+  'firestarter-fs9-or',
+  'flamberge-d',
+  'goshawk-ii-risc',
+  'grendel-i',
+  'griffin-c',
+  'griffin-grf-6s2',
+  'griffin-iic-9',
+  'grigori-c-grg-os-caelestis',
+  'gyrfalcon-4',
+  'ha-otoko-hr-unknown',
+  'hauptmann-ha1-ot',
+  'hellhound-8',
+  'hellion-g',
+  'hermes-ii-her-7a',
+  'highlander-hgn-732-jorgensson',
+  'inferno-inf-nor',
+  'jackalope-jlp-c',
+  'jackalope-jlp-kb',
+  'jackalope-jlp-nh-tepoztecatl',
+  'jade-hawk-jhk-03',
+  'jade-hawk-jhk-04',
+  'jagermech-jm7-dd',
+  'javelin-jvn-12n',
+  'juliano-jln-5a',
+  'juliano-jln-5c',
+  'kodiak-cale',
+  'koshi-z',
+  'locust-iic-9',
+  'mad-cat-blo',
+  'mad-cat-mk-iv-pr-risc',
+  'marauder-bounty-hunter-3138',
+  'marauder-c',
+  'minsk-2',
+  'minsk-mnk-101',
+  'mongrel-mgl-t1',
+  'mongrel-mgl-t2',
+  'morrigan-3',
+  'morrigan-4',
+  'naja-kto-19b-ec',
+  'nova-cat-i',
+  'orion-c',
+  'orion-c-2',
+  'phantom-r',
+  'pillager-plg-6z',
+  'pouncer-x',
+  'prowler-pwr-1x1',
+  'regent-c',
+  'ryoken-iii-c',
+  'ryoken-iii-xp-c',
+  'scarecrow-ucu-f4',
+  'shadow-cat-f',
+  'stalker-stk-9f',
+  'stealth-sth-5x',
+  'strider-sr1-oh',
+  'strider-sr1-or',
+  'sunder-sd1-of',
+  'sunder-sd1-og', // Clan chassis (same as SD1-OF/SD1-OR); +6 overcalc from other source
+  'sunder-sd1-or',
+  'templar-tlr1-or',
+  'templar-iii-tlr2-od',
+  'thresher-edward',
+  'thug-thg-11e-reich',
+  'thunderbolt-tdr-12r',
+  'thunderbolt-tdr-9w',
+  'trebuchet-tbt-9n',
+  'trebuchet-tbt-9r',
+  'uller-blo',
+  'valkyrie-vlk-qd5',
+  'vulture-mk-iii-blo',
+  'wendigo-xp-prime',
+  'xanthos-xnt-7o',
+]);
+
+// Units excluded from validation because they are not valid setups for the system
